rough working prototype website with database integration; lacking javascript and css
